+++ SlutprojektAPI/src/main/java/org/slutprojektapi/studentsCourses/StudentsCoursesController.java
@@ -20,14 +20,27 @@ public class StudentsCoursesController {
         return new ResponseEntity<>(studentsCoursesService.getAllStudentsCourses(), HttpStatus.OK);
     }
     @PostMapping(value = "/addStudentCourse")
-    public ResponseEntity<List<StudentsCoursesDTO>> createStudentCourse(@RequestParam(value = "studentId")Long studentId,
-                                                      @RequestParam(value = "courseId") Long courseId, StudentsCourses studentsCourse){
-        return studentsCoursesService.saveStudentCourse(studentsCourse,studentId,courseId);
+    public ResponseEntity<List<StudentsCoursesDTO>> createStudentCourse(@RequestParam(value = "studentId",required = false)Long studentId,
+                                                      @RequestParam(value = "courseId",required = false) Long courseId, StudentsCourses studentsCourse){
+        List<StudentsCoursesDTO> studentsCoursesDTOS = studentsCoursesService.saveStudentCourse(studentsCourse,studentId,courseId);
+        if (studentId<=0||courseId<=0){
+            return new ResponseEntity<>(HttpStatus.NOT_ACCEPTABLE);
+        } else {
+            return new ResponseEntity<>(studentsCoursesDTOS,HttpStatus.CREATED);
+        }
     }
     @PostMapping(value = "/removeStudentCourse")
-    public ResponseEntity<List<StudentsCoursesDTO>> removeStudentCourse(@RequestParam(value = "studentCourseId")Long studentCourseId){
-        studentsCoursesService.removeStudentCourse(studentCourseId);
-        return new ResponseEntity<>(studentsCoursesService.getAllStudentsCourses(),HttpStatus.OK);
+    public ResponseEntity<List<StudentsCoursesDTO>> removeStudentCourse(@RequestParam(value = "studentCourseId",required = false)Long studentCourseId){
+        if (studentCourseId<=0) {
+            return new ResponseEntity<>(HttpStatus.NOT_ACCEPTABLE);
+        }
+        if (studentCourseId.describeConstable().isEmpty()) {
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        } else {
+            studentsCoursesService.removeStudentCourse(studentCourseId);
+            List<StudentsCoursesDTO> studentsCoursesDTOS = studentsCoursesService.getAllStudentsCourses();
+            return new ResponseEntity<>(studentsCoursesDTOS,HttpStatus.OK);
+        }
     }
 
 }
